separated jsons for order activity and events, refactored creation of items object and included mapping for it
