+++ v0/destinations/drip/config.js
@@ -5,7 +5,9 @@ const ENDPOINT = "https://api.getdrip.com";
 const CONFIG_CATEGORIES = {
   IDENTIFY: { type: "identify", name: "DripIdentify" },
   TRACK: { type: "track", name: "DripTrack" },
-  CAMPAIGN: { type: "campaign", name: "DripCampaign" }
+  CAMPAIGN: { type: "campaign", name: "DripCampaign" },
+  ECOM: { type: "ecom", name: "DripEcom" },
+  PRODUCT: { type: "product", name: "DripProduct" }
 };
 
 const IDENTIFY_EXCLUSION_FIELDS = [
@@ -44,6 +46,15 @@ const TRACKING_EXLCUSION_FIELDS = [
   "id"
 ];
 
+const ecomEvents = [
+  "order updated",
+  "order completed",
+  "order refunded",
+  "order cancelled",
+  "checkout started",
+  "fulfilled"
+];
+
 const MAPPING_CONFIG = getMappingConfig(CONFIG_CATEGORIES, __dirname);
 
 module.exports = {
@@ -51,6 +62,9 @@ module.exports = {
   identifyMapping: MAPPING_CONFIG[CONFIG_CATEGORIES.IDENTIFY.name],
   trackMapping: MAPPING_CONFIG[CONFIG_CATEGORIES.TRACK.name],
   campaignMapping: MAPPING_CONFIG[CONFIG_CATEGORIES.CAMPAIGN.name],
+  ecomMapping: MAPPING_CONFIG[CONFIG_CATEGORIES.ECOM.name],
+  productMapping: MAPPING_CONFIG[CONFIG_CATEGORIES.PRODUCT.name],
   IDENTIFY_EXCLUSION_FIELDS,
-  TRACKING_EXLCUSION_FIELDS
+  TRACKING_EXLCUSION_FIELDS,
+  ecomEvents
 };
